# refactor: use StringBuilder instead of List