# Replace `List<string>` with `StringBuilder` in `Htmlizer`
The PR replaces the current `List<string>` with a `StringBuilder`.  This brings mainly 3 advantages:
 * Make the intent better visible
 * Less allocations
 * Better performance

I know the original implementation comes from @SteveSandersonMS, but not sure what the motivation was.
Anyway I thought I take the lead and propose a different approach with the given advantages on the top.
Feedback very welcome.

# Rendering a sample component
The test just renders a rather small component. The results will be even better when we render bigger components.
For the test I used the `MultipleChildContent` as reference. 

## Old
```csharp
|          Method |     Mean |   Error |   StdDev |   Gen 0 |   Gen 1 | Allocated |
|---------------- |---------:|--------:|---------:|--------:|--------:|----------:|
| RenderComponent | 388.9 us | 7.72 us | 19.07 us | 42.9688 | 14.1602 |    179 KB |
```

## New
```csharp
|          Method |     Mean |   Error |   StdDev |   Gen 0 |   Gen 1 | Allocated |
|---------------- |---------:|--------:|---------:|--------:|--------:|----------:|
| RenderComponent | 368.3 us | 7.77 us | 22.03 us | 41.0156 | 13.6719 |    170 KB |
```

# General
In general `StringBuilder` is faster and uses less allocations than a `List<string>`.
I built a small [sample benchmark](https://gist.github.com/linkdotnet/fc8ef970b504a6c69f633053383c8084) with the following results:
```csharp
|          Method |       Mean |    Error |   StdDev |     Median | Ratio | RatioSD |  Gen 0 | Allocated |
|---------------- |-----------:|---------:|---------:|-----------:|------:|--------:|-------:|----------:|
|          ByList | 1,993.5 ns | 58.49 ns | 167.8 ns | 1,988.0 ns |  1.00 |    0.00 | 0.7744 |      3 KB |
| ByStringBuilder |   896.6 ns | 78.55 ns | 231.6 ns |   811.9 ns |  0.45 |    0.12 | 0.5980 |      2 KB |

```