@egil 2022-05-03
Great idea, really like that you find these relatively low hanging fruits. Lets use a StringBuilder pool for even better performance (https://www.meziantou.net/stringbuilder-performance-pitfalls.htm#use-a-pool-of-string).

Ps. I have been meaning to set up it would be a great idea to set up micro performance tests to (benchmark.net) to start monitoring the speed of bUnit, and potentially in the future, try to optimize it. Looks like you have something set already here.