@linkdotnet 2022-05-03
Good idea. I will create a separate PR for the setup... that was rather quick and dirty and should not be commited at all :D

Anyway I added the `StringBuilderPool`.