@egil 2022-05-03
Might be interesting to do some benchmarking to see if an object pool is always better. Perhaps it only really matter if a TestContext is used with 20-30 re-renders.