@linkdotnet 2022-05-03
I can do some testing later today or tomorrow to verify that.
Maybe it would even be worth to even tackle #703 to some extend with this PR as well... what do you think?