@linkdotnet 2022-05-03
One thing I noticed when running the benchmarks, if we "reuse" the `TestContext` like this:
```csharp
public void Benchmark
{
    private readonly TestContext context = new();
    
    [Benchmark]
    public IRenderFragmentBase Render() => context.RenderComponent<Component>();
}
```

It becomes slower and slower and slower. Only if I create the `TestContext` and dipose it directly after use the runtime is const. 
You can super easily reproduce this via:
```csharp
var ctx = new TestContext();

var stopwatch = Stopwatch.StartNew();
for (var i = 0; i < 100_000; i++)
{
    _ = ctx.RenderComponent<Counter>();

    if (i % 1000 == 0)
    {
        Console.WriteLine($"Average run took: {stopwatch.ElapsedMilliseconds / 1000} ms");
        stopwatch.Restart();
    }
}
```

Output for me:
> Average run took: 0 ms
Average run took: 0 ms
Average run took: 2 ms
Average run took: 4 ms
Average run took: 6 ms
Average run took: 9 ms
Average run took: 10 ms
Average run took: 13 ms
Average run took: 15 ms
Average run took: 19 ms
Average run took: 19 ms
Average run took: 20 ms

I render here multiple 1000 components, so that is something normal users will most likely never experience, but when we want to benchmark certain aspect we might want to investigate what is pilling up here.