@egil 2022-05-03
I think it makes sense that it becomes slower, since there is some internal dictionaries/arrays in the TestRenderer and Renderer that grows whenever we render a component. I am surprised. But it might be worth bringing up in the docs as another good reason not to reuse the TestContext between tests (ive seen people do this).