@linkdotnet 2022-05-03
I checked with some benchmark. The `StringBuilderPool` causes a bit more memory with almost the same performance.
I guess we could drop the pooling.. or maybe I did a major flaw with those few lines of code.