@egil 2022-05-03
Its a balancing act between maintainable code and perf. Fine to drop StringBuilderPool if there is little benefit at the typical usage pattern. `ValueStringBuilder `sounds interesting. If its an easy thing to try, it would be fun to do.