@linkdotnet 2022-05-03
The last commit adds the `ValueStringBuilder`. I took only the necessary parts from my [own library here](https://github.com/linkdotnet/StringBuilder).

The results are pretty nice. Again I use the `MultipleChildContent`as reference.
The test looks like this:
```csharp
[Benchmark]
public IRenderedFragmentBase RenderCounter()
{
    using var ctx = new TestContext();
    return ctx.RenderComponent<MultipleChildContent>();
}
```

## Using `List<T>`
```csharp
|          Method |     Mean |   Error |   StdDev |   Gen 0 |   Gen 1 | Allocated |
|---------------- |---------:|--------:|---------:|--------:|--------:|----------:|
| RenderComponent | 388.9 us | 7.72 us | 19.07 us | 42.9688 | 14.1602 |    179 KB |
```
## Using `StringBuilder`
```csharp
|          Method |     Mean |   Error |   StdDev |   Gen 0 |   Gen 1 | Allocated |
|---------------- |---------:|--------:|---------:|--------:|--------:|----------:|
| RenderComponent | 368.3 us | 7.77 us | 22.03 us | 41.0156 | 13.6719 |    170 KB |
```

## Using `ValueStringtBuilder`
```csharp
|        Method |     Mean |   Error |   StdDev |   Median |   Gen 0 |   Gen 1 | Allocated |
|-------------- |---------:|--------:|---------:|---------:|--------:|--------:|----------:|
| RenderCounter | 309.9 us | 8.97 us | 25.45 us | 301.0 us | 40.5273 | 13.1836 |    166 KB |
```

Now the major downside of a `ValueStringBuilder` is the fact that it is a `ref struct` which forces passing the object around a lot via `ref`.