@linkdotnet 2022-05-03
Seems like some tests are failing. I'll investigate that later. From my point of view the "normal" `StringBuilder` would be a nice middleground. Basically adds no more complexity than the solution before and we get some performance improvements. The `ValueStringBuilder` is quite deep into the code.