@egil 2022-05-03
Agreed, lets stick with a regular `StringBuilder`. Good work btw.!